Step3(refactor): app module 패키지 정리
