+++ app/src/main/java/camp/nextstep/edu/github/Mappers.kt
@@ -2,6 +2,7 @@ package camp.nextstep.edu.github
 
 import camp.nextstep.edu.github.domain.model.GitHubRepo
 import camp.nextstep.edu.github.domain.model.GitHubRepos
+import camp.nextstep.edu.github.model.GitHubRepoViewData
 
 
 fun GitHubRepos.toViewDataList() = repositories.map { it.toViewData() }
