+++ app/src/main/java/camp/nextstep/edu/github/model/GitHubRepoViewData.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.github
+package camp.nextstep.edu.github.model
 
 
 data class GitHubRepoViewData(val id: Long, val authorRepoStr: String, val description: String)
