+++ app/src/main/java/camp/nextstep/edu/github/ui/MainActivity.kt
@@ -1,10 +1,13 @@
-package camp.nextstep.edu.github
+package camp.nextstep.edu.github.ui
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
+import camp.nextstep.edu.github.GitHubReposAdapter
+import camp.nextstep.edu.github.viewmodel.MainViewModel
 import camp.nextstep.edu.github.databinding.ActivityMainBinding
+import camp.nextstep.edu.github.toViewDataList
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
